pyproject.toml
[project]
name = "discovery_piscine"
version = "0.1.0"
description = "42 Firenze - Discovery Piscine Python"
authors = [{ name = "Julio Formiga", email = "julio.formiga@gmail.com" }]
dependencies = []
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }

[tool.pytest.ini_options]
pythonpaths = ["src"]
addopts = "-p no:warnings"

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = ["pytest-cov>=5.0.0", "pytest-sugar>=1.0.0"]

[tool.poetry]
name = "42-piscine-python"
version = "0.1.0"
description = "42 Firenze - Discovery Python"
authors = ["Julio Formiga <julio.formiga@gmail.com>"]
license = "MIT"
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest-cov = "^5.0.0"
pytest-sugar = "^1.0.0"
